requirements.txt
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0
pandas>=2.0.0 